autoresearch-master/requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
Pillow>=10.0
pandas>=2.0
scikit-learn>=1.3
numpy>=2.2
matplotlib>=3.10
tqdm>=4.65
